Revert "attempting to move copy code around"
This reverts commit a868f2862cca186a86bff2726402332795aa6fa6.
