+++ src/main/resources/edu/cornell/kfs/coa/document/workflow/ObjectCodeActivationGlobalMaintenanceDocument.xml
@@ -6,7 +6,6 @@
       <parent>COA</parent>
       <label>Object Code Activation Global</label>
       <authorizer>org.kuali.kfs.krad.workflow.authorizer.CfDocumentTypeAuthorizer</authorizer>
-      <docHandler>${kfs.url}/kr/objectcodeactivationglobalmaintenance.do?methodToCall=docHandler</docHandler>
       <active>true</active>
       <routingVersion>1</routingVersion>
       <routePaths>
