+++ src/main/webapp/WEB-INF/institutional-struts-config.xml
@@ -221,12 +221,6 @@
             <forward name="basic" path="/jsp/fp/CorporateBilledCorporatePaid.jsp"/>
         </action>
         
-        <action path="/objectcodeactivationglobalmaintenance" name="KualiMaintenanceForm" attribute="KualiForm" scope="request"
-                parameter="methodToCall" validate="false" input="/WEB-INF/jsp/KualiMaintenanceDocument.jsp"
-                type="edu.cornell.kfs.coa.document.web.struts.CuObjectCodeActivationGlobalMaintenanceDocumentAction">
-	        <forward name="basic" path="/WEB-INF/jsp/KualiMaintenanceDocument.jsp" />
-	    </action>
-        
     </action-mappings>
     
 </struts-config>
